+++ jib-maven-plugin/src/test/java/com/google/cloud/tools/jib/maven/skaffold/FilesMojoV2Test.java
@@ -153,7 +153,6 @@ public class FilesMojoV2Test {
             complexServiceRoot.resolve("src/main/resources2").toString(),
             complexServiceRoot.resolve("src/main/jib1").toString(),
             complexServiceRoot.resolve("src/main/jib2").toString(),
-            Paths.get("/").toAbsolutePath().resolve("some/random/absolute/path/jib3").toString(),
             // this test expects standard .m2 locations
             Paths.get(
                     System.getProperty("user.home"),