+++ jib-plugins-common/src/test/java/com/google/cloud/tools/jib/plugins/common/PluginConfigurationProcessorTest.java
@@ -199,7 +199,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     ContainerBuildPlan buildPlan = processCommonConfiguration();
 
     assertThat(buildPlan.getEntrypoint())
@@ -220,7 +221,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, IOException, IncompatibleBaseImageJavaVersionException,
           InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidImageReferenceException, NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     Path extraDirectory = Paths.get(Resources.getResource("core/layer").toURI());
     Mockito.<List<?>>when(rawConfiguration.getExtraDirectories())
         .thenReturn(Arrays.asList(new TestExtraDirectoriesConfiguration(extraDirectory)));
@@ -260,12 +262,25 @@ public class PluginConfigurationProcessorTest {
   }
 
   @Test
+  public void testPluginConfigurationProcessor__errorOnExtraDirectoryPathNotFound()
+      throws URISyntaxException, NumberFormatException {
+    Path extraDirectory = Paths.get(Resources.getResource("core/layer").toURI()).resolve("xyz");
+    Mockito.<List<?>>when(rawConfiguration.getExtraDirectories())
+        .thenReturn(Arrays.asList(new TestExtraDirectoriesConfiguration(extraDirectory)));
+
+    Exception exception =
+        assertThrows(ExtraDirectoryNotFoundException.class, this::processCommonConfiguration);
+    assertThat(exception).hasMessageThat().isEqualTo(extraDirectory.toString());
+  }
+
+  @Test
   public void testPluginConfigurationProcessor_cacheDirectorySystemProperties()
       throws InvalidContainerVolumeException, MainClassInferenceException, InvalidAppRootException,
           IOException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidImageReferenceException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     System.setProperty(PropertyNames.BASE_IMAGE_CACHE, "new/base/cache");
     System.setProperty(PropertyNames.APPLICATION_CACHE, "/new/application/cache");
 
@@ -340,7 +355,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getEntrypoint())
         .thenReturn(Optional.of(Arrays.asList("custom", "entrypoint")));
 
@@ -446,7 +462,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(projectProperties.isWarProject()).thenReturn(true);
 
     ContainerBuildPlan buildPlan = processCommonConfiguration();
@@ -463,7 +480,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(projectProperties.isWarProject()).thenReturn(false);
 
     ContainerBuildPlan buildPlan = processCommonConfiguration();
@@ -483,7 +501,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getExtraClasspath()).thenReturn(Collections.singletonList("/foo"));
     when(projectProperties.isWarProject()).thenReturn(false);
 
@@ -504,7 +523,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, IOException, InvalidWorkingDirectoryException,
           InvalidPlatformException, InvalidContainerVolumeException,
           IncompatibleBaseImageJavaVersionException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getExtraClasspath()).thenReturn(Collections.singletonList("/foo"));
     when(projectProperties.getMajorJavaVersion()).thenReturn(9);
 
@@ -543,7 +563,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, IOException, InvalidWorkingDirectoryException,
           InvalidPlatformException, InvalidContainerVolumeException,
           IncompatibleBaseImageJavaVersionException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getMainClass()).thenReturn(Optional.of("invalid main class"));
     when(rawConfiguration.getEntrypoint()).thenReturn(Optional.of(Arrays.asList("bash")));
 
@@ -579,7 +600,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getUser()).thenReturn(Optional.of("customUser"));
 
     ContainerBuildPlan buildPlan = processCommonConfiguration();
@@ -592,7 +614,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     ContainerBuildPlan buildPlan = processCommonConfiguration();
     assertThat(buildPlan.getUser()).isNull();
   }
@@ -603,7 +626,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getEntrypoint())
         .thenReturn(Optional.of(Arrays.asList("custom", "entrypoint")));
     when(rawConfiguration.getJvmFlags()).thenReturn(Collections.singletonList("jvmFlag"));
@@ -624,7 +648,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getEntrypoint())
         .thenReturn(Optional.of(Arrays.asList("custom", "entrypoint")));
     when(rawConfiguration.getMainClass()).thenReturn(Optional.of("java.util.Object"));
@@ -645,7 +670,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getEntrypoint())
         .thenReturn(Optional.of(Arrays.asList("custom", "entrypoint")));
     when(rawConfiguration.getExpandClasspathDependencies()).thenReturn(true);
@@ -666,7 +692,7 @@ public class PluginConfigurationProcessorTest {
           IncompatibleBaseImageJavaVersionException, InvalidPlatformException,
           InvalidContainerVolumeException, MainClassInferenceException, InvalidAppRootException,
           InvalidWorkingDirectoryException, InvalidFilesModificationTimeException,
-          InvalidContainerizingModeException {
+          InvalidContainerizingModeException, ExtraDirectoryNotFoundException {
     when(rawConfiguration.getMainClass()).thenReturn(Optional.of("java.util.Object"));
     when(projectProperties.isWarProject()).thenReturn(true);
 
@@ -688,7 +714,7 @@ public class PluginConfigurationProcessorTest {
           IncompatibleBaseImageJavaVersionException, InvalidPlatformException,
           InvalidContainerVolumeException, MainClassInferenceException, InvalidAppRootException,
           InvalidWorkingDirectoryException, InvalidFilesModificationTimeException,
-          InvalidContainerizingModeException {
+          InvalidContainerizingModeException, ExtraDirectoryNotFoundException {
     when(rawConfiguration.getExpandClasspathDependencies()).thenReturn(true);
     when(projectProperties.isWarProject()).thenReturn(true);
 
@@ -710,7 +736,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(rawConfiguration.getAppRoot()).thenReturn("/my/app");
 
     ContainerBuildPlan buildPlan = processCommonConfiguration();
@@ -727,7 +754,8 @@ public class PluginConfigurationProcessorTest {
           InvalidAppRootException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     when(projectProperties.isWarProject()).thenReturn(true);
     when(rawConfiguration.getFromImage()).thenReturn(Optional.of("custom-base-image"));
 
@@ -1167,7 +1195,8 @@ public class PluginConfigurationProcessorTest {
           IOException, InvalidWorkingDirectoryException, InvalidPlatformException,
           InvalidContainerVolumeException, IncompatibleBaseImageJavaVersionException,
           NumberFormatException, InvalidContainerizingModeException,
-          InvalidFilesModificationTimeException, InvalidCreationTimeException {
+          InvalidFilesModificationTimeException, InvalidCreationTimeException,
+          ExtraDirectoryNotFoundException {
     JibContainerBuilder containerBuilder =
         PluginConfigurationProcessor.processCommonConfiguration(
             rawConfiguration, ignored -> Optional.empty(), projectProperties, containerizer);